Finished implementation of API docs generation
This also required changing the functionality from the last commit c2a0c31e9bcc8bc400fd39ff2933364905789476, as the building process was so slow and memory-heavy that it was not possible to build it effectively. As such, a new memory strategy was required to optimise the process and also more effectively use the given constraints. This means that the new API generator will, instead of building using Parcel, use an existing smaller build of the static docs and reuse the compiled files as a way to save memory & performance. View function 'injectMarkdownIntoTemplate()' for more info on the injection process.