Implemented phone-optimised design for the playground
Minor bugs still exist due to resizing. Additionally, in case the version tag is too long, the version will be cut off and made a scrollable element. This should be made into an animation, but with this commit that will be left out to not break the existing SCSS code.